feat(jade): 添加智能方块放置器的Jade插件支持
- 新增SmartBlockPlacerProvider和SmartBlockPlacerClientProvider实现
- 在AnvilCraftJadePlugin中注册智能方块放置器的数据提供者和组件提供者
- 完善相关语言文件，添加智能方块放置器的配置和提示文本
- 优化渲染基类BaseFluidHandlerHolderRenderer，修复流体着色获取方式
- 为多种流体模型统一添加默认白色着色处理
- 移除SmartBlockPlacerProvider中客户端相关代码，转移至客户端实现类
- 服务器端通过NBT同步方块实体数据，无需额外服务器数据传输
