+++ src/main/java/dev/dubhe/anvilcraft/data/lang/JadeLang.java
@@ -33,5 +33,18 @@ public class JadeLang {
         provider.add("tooltip.anvilcraft.burning_heater.jade.can_smelt", "Can Smelt: %s");
         provider.add("tooltip.anvilcraft.burning_heater.jade.can_smelt.yes", "Yes");
         provider.add("tooltip.anvilcraft.burning_heater.jade.can_smelt.no", "No");
+
+        provider.add("config.jade.plugin_anvilcraft.smart_block_placer_provider", "Smart Block Placer");
+
+        provider.add("tooltip.anvilcraft.smart_block_placer.jade.operation_mode", "Operation Mode: %s");
+        provider.add("tooltip.anvilcraft.smart_block_placer.jade.mode.normal", "Normal");
+        provider.add("tooltip.anvilcraft.smart_block_placer.jade.mode.blueprint", "Blueprint");
+        provider.add("tooltip.anvilcraft.smart_block_placer.jade.placement_mode", "Placement Mode: %s");
+        provider.add("tooltip.anvilcraft.smart_block_placer.jade.placement.pickup", "Pickup");
+        provider.add("tooltip.anvilcraft.smart_block_placer.jade.placement.move", "Move");
+        provider.add("tooltip.anvilcraft.smart_block_placer.jade.blueprint_name", "Blueprint: %s");
+        provider.add("tooltip.anvilcraft.smart_block_placer.jade.missing_mode", "Missing Block: %s");
+        provider.add("tooltip.anvilcraft.smart_block_placer.jade.missing.skip", "Skip");
+        provider.add("tooltip.anvilcraft.smart_block_placer.jade.missing.stop", "Stop");
     }
 }
